feat: Consolidate search module into feature
