+++ feature/notification/build.gradle.kts
@@ -8,4 +8,6 @@ android {
 
 dependencies {
     implementation(libs.androidx.navigation.compose)
+
+    implementation(projects.core.domain)
 }
