--- notification/data/build.gradle.kts
@@ -1,16 +1,0 @@
-plugins {
-    alias(libs.plugins.hyunjung.cherrydan.android.library)
-    alias(libs.plugins.hyunjung.cherrydan.jvm.ktor)
-}
-
-android {
-    namespace = "com.hyunjung.notification.data"
-}
-
-dependencies {
-    implementation(libs.bundles.koin)
-
-    implementation(projects.notification.domain)
-    implementation(projects.core.domain)
-    implementation(projects.core.data)
-}
