--- search/domain/build.gradle.kts
@@ -1,7 +1,0 @@
-plugins {
-    alias(libs.plugins.hyunjung.cherrydan.jvm.library)
-}
-
-dependencies {
-    implementation(projects.core.domain)
-}
