--- search/domain/src/main/java/com/hyunjung/search/domain/MyClass.kt
@@ -1,4 +1,0 @@
-package com.hyunjung.search.domain
-
-class MyClass {
-}
