--- search/presentation/build.gradle.kts
@@ -1,14 +1,0 @@
-plugins {
-    alias(libs.plugins.hyunjung.cherrydan.android.feature.ui)
-}
-
-android {
-    namespace = "com.hyunjung.search.presentation"
-}
-
-dependencies {
-    implementation(libs.androidx.navigation.compose)
-
-    implementation(projects.core.domain)
-    implementation(projects.search.domain)
-}
