+++ settings.gradle.kts
@@ -30,10 +30,8 @@ include(":core:domain")
 include(":core:data")
 include(":core:database")
 include(":core:network")
-include(":search:presentation")
-include(":search:domain")
-include(":search:data")
 include(":core:model")
 include(":feature:auth")
 include(":feature:home")
 include(":feature:notification")
+include(":feature:search")
